+++ dump/plugins/portchannel.py
@@ -11,7 +11,6 @@ class Portchannel(Executor):
 
     def __init__(self, match_engine=None):
         super().__init__(match_engine)
-        self.ret_temp = {}
         self.ns = ''
         self.lag_members = set()
 
@@ -37,14 +36,6 @@ class Portchannel(Executor):
         lag_type_objs_asic = self.init_lag_member_type_obj_asic_info()
         self.init_lag_asic_info(lag_type_objs_asic)
         return self.ret_temp
-
-    def add_to_ret_template(self, table, db, keys, err):
-        if not err and keys:
-            self.ret_temp[db]["keys"].extend(keys)
-            return True
-        else:
-            self.ret_temp[db]["tables_not_found"].extend([table])
-            return False
 
     def init_lag_config_info(self):
         req = MatchRequest(db="CONFIG_DB", table="PORTCHANNEL", key_pattern=self.lag_name, ns=self.ns)
