+++ dump/plugins/portchannel_member.py
@@ -10,7 +10,6 @@ class Portchannel_Member(Executor):
 
     def __init__(self, match_engine=None):
         super().__init__(match_engine)
-        self.ret_temp = {}
         self.ns = ''
         self.lag_member_key = ''
         self.lag = ''
@@ -38,14 +37,6 @@ class Portchannel_Member(Executor):
         # ASIC_DB
         self.init_lag_member_type_obj_asic_info()
         return self.ret_temp
-
-    def add_to_ret_template(self, table, db, keys, err):
-        if not err and keys:
-            self.ret_temp[db]["keys"].extend(keys)
-            return True
-        else:
-            self.ret_temp[db]["tables_not_found"].extend([table])
-            return False
 
     def init_lag_member_config_info(self):
         req = MatchRequest(db="CONFIG_DB", table="PORTCHANNEL_MEMBER", key_pattern=self.lag_member_key, ns=self.ns)
