+++ dump/plugins/vxlan_tunnel_map.py
@@ -11,7 +11,6 @@ class Vxlan_tunnel_map(Executor):
 
     def __init__(self, match_engine=None):
         super().__init__(match_engine)
-        self.ret_temp = {}
         self.ns = ''
         self.vlan = ''
         self.vni = ''
@@ -34,14 +33,6 @@ class Vxlan_tunnel_map(Executor):
         self.init_asic_vxlan_tunnel_map_entry_info()
         self.init_asic_vxlan_tunnel_map_info()
         return self.ret_temp
-
-    def add_to_ret_template(self, table, db, keys, err):
-        if not err and keys:
-            self.ret_temp[db]["keys"].extend(keys)
-            return True
-        else:
-            self.ret_temp[db]["tables_not_found"].extend([table])
-            return False
 
     def init_vxlan_tunnel_map_config_info(self, vxlan_tunnel_map_name):
         req = MatchRequest(db="CONFIG_DB", table="VXLAN_TUNNEL_MAP", key_pattern=vxlan_tunnel_map_name, ns=self.ns,
