+++ tests/dump_tests/module_tests/evpn_test.py
@@ -2,7 +2,7 @@ import json
 import os
 import pytest
 from deepdiff import DeepDiff
-from dump.helper import create_template_dict
+from dump.helper import create_template_dict, populate_mock
 from dump.plugins.evpn import Evpn
 from dump.match_infra import MatchEngine, ConnectionPool
 from swsscommon.swsscommon import SonicV2Connector
@@ -21,18 +21,6 @@ dedicated_dbs['ASIC_DB'] = os.path.join(evpn_files_path, "asic_db.json")
 dedicated_dbs['STATE_DB'] = os.path.join(evpn_files_path, "state_db.json")
 
 
-def populate_mock(db, db_names):
-    for db_name in db_names:
-        db.connect(db_name)
-        # Delete any default data
-        db.delete_all_by_pattern(db_name, "*")
-        with open(dedicated_dbs[db_name]) as f:
-            mock_json = json.load(f)
-        for key in mock_json:
-            for field, value in mock_json[key].items():
-                db.set(db_name, key, field, value)
-
-
 @pytest.fixture(scope="class", autouse=True)
 def match_engine():
 
@@ -46,7 +34,7 @@ def match_engine():
     # popualate the db with mock data
     db_names = list(dedicated_dbs.keys())
     try:
-        populate_mock(db, db_names)
+        populate_mock(db, db_names, dedicated_dbs)
     except Exception as e:
         assert False, "Mock initialization failed: " + str(e)
