+++ tests/dump_tests/module_tests/portchannel_test.py
@@ -4,7 +4,7 @@ import unittest
 import pytest
 from deepdiff import DeepDiff
 from mock import patch
-from dump.helper import create_template_dict, sort_lists
+from dump.helper import create_template_dict, sort_lists, populate_mock
 from dump.plugins.portchannel import Portchannel
 from dump.match_infra import MatchEngine, ConnectionPool
 from swsscommon.swsscommon import SonicV2Connector
@@ -25,18 +25,6 @@ dedicated_dbs['ASIC_DB'] = os.path.join(port_files_path, "asic_db.json")
 dedicated_dbs['STATE_DB'] = os.path.join(port_files_path, "state_db.json")
 
 
-def populate_mock(db, db_names):
-    for db_name in db_names:
-        db.connect(db_name)
-        # Delete any default data
-        db.delete_all_by_pattern(db_name, "*")
-        with open(dedicated_dbs[db_name]) as f:
-            mock_json = json.load(f)
-        for key in mock_json:
-            for field, value in mock_json[key].items():
-                db.set(db_name, key, field, value)
-
-
 @pytest.fixture(scope="class", autouse=True)
 def match_engine():
 
@@ -50,7 +38,7 @@ def match_engine():
     # popualate the db with mock data
     db_names = list(dedicated_dbs.keys())
     try:
-        populate_mock(db, db_names)
+        populate_mock(db, db_names, dedicated_dbs)
     except Exception as e:
         assert False, "Mock initialization failed: " + str(e)
