+++ tests/dump_tests/module_tests/route_test.py
@@ -4,7 +4,7 @@ import pytest
 import json
 from deepdiff import DeepDiff
 from mock import patch
-from dump.helper import create_template_dict
+from dump.helper import create_template_dict, populate_mock
 from dump.plugins.route import Route
 from dump.match_infra import MatchEngine, ConnectionPool
 from swsscommon.swsscommon import SonicV2Connector
@@ -24,18 +24,6 @@ dedicated_dbs['APPL_DB'] = os.path.join(Route_files_path, "appl_db.json")
 dedicated_dbs['ASIC_DB'] = os.path.join(Route_files_path, "asic_db.json")
 
 
-def populate_mock(db, db_names):
-    for db_name in db_names:
-        db.connect(db_name)
-        # Delete any default data
-        db.delete_all_by_pattern(db_name, "*")
-        with open(dedicated_dbs[db_name]) as f:
-            mock_json = json.load(f)
-        for key in mock_json:
-            for field, value in mock_json[key].items():
-                db.set(db_name, key, field, value)
-
-
 @pytest.fixture(scope="class", autouse=True)
 def match_engine():
 
@@ -49,7 +37,7 @@ def match_engine():
     # popualate the db with mock data
     db_names = list(dedicated_dbs.keys())
     try:
-        populate_mock(db, db_names)
+        populate_mock(db, db_names, dedicated_dbs)
     except Exception as e:
         assert False, "Mock initialization failed: " + str(e)
